Add security note regarding availability of GITHUB_TOKEN
There is a chance for third parties to be able to upload arbitrary files to your release page by sending you pull requests etc. which print out the variable value.

Also, it's a really simple way to prevent uploading files to continuous releases on branches other than `master`. If the credentials are simply not available, uploadtool cannot upload anything, right?